+++ ipwhois/ipwhois.py
@@ -36,14 +36,15 @@ class IPWhois:
     IPv4 and IPv6 addresses.
 
     Args:
-        address: An IPv4 or IPv6 address as a string, integer, IPv4Address, or
-            IPv6Address.
-        timeout: The default timeout for socket connections in seconds.
-        proxy_opener: The urllib.request.OpenerDirector request for proxy
-            support or None.
-        allow_permutations: allow net.Net() to use additional methods if DNS
-            lookups to Cymru fail. *WARNING* deprecated in favor of new
-            argument asn_methods.
+        address (:obj:`str`/:obj:`int`/:obj:`IPv4Address`/:obj:`IPv6Address`):
+            An IPv4 or IPv6 address
+        timeout (:obj:`int`): The default timeout for socket connections in
+            seconds. Defaults to 5.
+        proxy_opener (:obj:`urllib.request.OpenerDirector`): The request for
+            proxy support. Defaults to None.
+        allow_permutations (:obj:`bool`): Allow net.Net() to use additional
+            methods if DNS lookups to Cymru fail. *WARNING* deprecated in
+            favor of new argument asn_methods. Defaults to True.
     """
 
     def __init__(self, address, timeout=5, proxy_opener=None,
@@ -68,19 +69,6 @@ class IPWhois:
             self.address_str, str(self.timeout), repr(self.net.opener)
         )
 
-    def lookup(self, *args, **kwargs):
-        """
-        Temporary wrapper for legacy whois lookups (moved to
-        IPWhois.lookup_whois()). This will be removed in a future
-        release (1.0.0).
-        """
-
-        from warnings import warn
-        warn('IPWhois.lookup() has been deprecated and will be removed. '
-             'You should now use IPWhois.lookup_whois() for legacy whois '
-             'lookups.')
-        return self.lookup_whois(*args, **kwargs)
-
     def lookup_whois(self, inc_raw=False, retry_count=3, get_referral=False,
                      extra_blacklist=None, ignore_referral_errors=False,
                      field_list=None, asn_alts=None, extra_org_map=None,
@@ -91,63 +79,77 @@ class IPWhois:
         address via port 43 (WHOIS).
 
         Args:
-            inc_raw: Boolean for whether to include the raw whois results in
-                the returned dictionary.
-            retry_count: The number of times to retry in case socket errors,
-                timeouts, connection resets, etc. are encountered.
-            get_referral: Boolean for whether to retrieve referral whois
-                information, if available.
-            extra_blacklist: A list of blacklisted whois servers in addition to
-                the global BLACKLIST.
-            ignore_referral_errors: Boolean for whether to ignore and continue
-                when an exception is encountered on referral whois lookups.
-            field_list: If provided, a list of fields to parse:
+            inc_raw (:obj:`bool`): Whether to include the raw whois results in
+                the returned dictionary. Defaults to False.
+            retry_count (:obj:`int`): The number of times to retry in case
+                socket errors, timeouts, connection resets, etc. are
+                encountered. Defaults to 3.
+            get_referral (:obj:`bool`): Whether to retrieve referral whois
+                information, if available. Defaults to False.
+            extra_blacklist (:obj:`list`): Blacklisted whois servers in
+                addition to the global BLACKLIST. Defaults to None.
+            ignore_referral_errors (:obj:`bool`): Whether to ignore and
+                continue when an exception is encountered on referral whois
+                lookups. Defaults to False.
+            field_list (:obj:`list`): If provided, a list of fields to parse:
                 ['name', 'handle', 'description', 'country', 'state', 'city',
                 'address', 'postal_code', 'emails', 'created', 'updated']
-            asn_alts: List of additional lookup types to attempt if the
+                If None, defaults to all.
+            asn_alts (:obj:`list`): Additional lookup types to attempt if the
                 ASN dns lookup fails. Allow permutations must be enabled.
-                Defaults to all ['whois', 'http']. *WARNING* deprecated in
-                favor of new argument asn_methods.
-            extra_org_map: Dictionary mapping org handles to RIRs. This is for
-                limited cases where ARIN REST (ASN fallback HTTP lookup) does
-                not show an RIR as the org handle e.g., DNIC (which is now the
-                built in ORG_MAP) e.g., {'DNIC': 'arin'}. Valid RIR values are
-                (note the case-sensitive - this is meant to match the REST
-                result): 'ARIN', 'RIPE', 'apnic', 'lacnic', 'afrinic'
-            inc_nir: Boolean for whether to retrieve NIR (National Internet
+                If None, defaults to all ['whois', 'http']. *WARNING*
+                deprecated in favor of new argument asn_methods.
+            extra_org_map (:obj:`dict`): Dictionary mapping org handles to
+                RIRs. This is for limited cases where ARIN REST (ASN fallback
+                HTTP lookup) does not show an RIR as the org handle e.g., DNIC
+                (which is now the built in ORG_MAP) e.g., {'DNIC': 'arin'}.
+                Valid RIR values are (note the case-sensitive - this is meant
+                to match the REST result):
+                'ARIN', 'RIPE', 'apnic', 'lacnic', 'afrinic'
+                Defaults to None.
+            inc_nir (:obj:`bool`): Whether to retrieve NIR (National Internet
                 Registry) information, if registry is JPNIC (Japan) or KRNIC
                 (Korea). If True, extra network requests will be required.
                 If False, the information returned for JP or KR IPs is
-                severely restricted.
-            nir_field_list: If provided and inc_nir, a list of fields to parse:
+                severely restricted. Defaults to True.
+            nir_field_list (:obj:`list`): If provided and inc_nir, a list of
+                fields to parse:
                 ['name', 'handle', 'country', 'address', 'postal_code',
                 'nameservers', 'created', 'updated', 'contacts']
-            asn_methods: List of ASN lookup types to attempt, in order.
-                Defaults to all ['dns', 'whois', 'http'].
-            get_asn_description: Boolean for whether to run an additional
+                If None, defaults to all.
+            asn_methods (:obj:`list`): ASN lookup types to attempt, in order.
+                If None, defaults to all ['dns', 'whois', 'http'].
+            get_asn_description (:obj:`bool`): Whether to run an additional
                 query when pulling ASN information via dns, in order to get
-                the ASN description.
+                the ASN description. Defaults to True.
 
         Returns:
-            Dictionary:
+            dict: The IP whois lookup results
 
-            :query: The IP address (String)
-            :asn: The Autonomous System Number (String)
-            :asn_date: The ASN Allocation date (String)
-            :asn_registry: The assigned ASN registry (String)
-            :asn_cidr: The assigned ASN CIDR (String)
-            :asn_country_code: The assigned ASN country code (String)
-            :asn_description: The ASN description (String)
-            :nets: Dictionaries containing network information which consists
-                of the fields listed in the ipwhois.whois.RIR_WHOIS dictionary.
-                (List)
-            :raw: Raw whois results if the inc_raw parameter is True. (String)
-            :referral: Dictionary of referral whois information if get_referral
-                is True and the server isn't blacklisted. Consists of fields
-                listed in the ipwhois.whois.RWHOIS dictionary.
-            :raw_referral: Raw referral whois results if the inc_raw parameter
-                is True. (String)
-            :nir: (Dictionary) - nir.NIRWhois() results if inc_nir is True.
+            ::
+
+                {
+                    'query' (str) - The IP address
+                    'asn' (str) - The Autonomous System Number
+                    'asn_date' (str) - The ASN Allocation date
+                    'asn_registry' (str) - The assigned ASN registry
+                    'asn_cidr' (str) - The assigned ASN CIDR
+                    'asn_country_code' (str) - The assigned ASN country code
+                    'asn_description' (str) - The ASN description
+                    'nets' (list) - Dictionaries containing network
+                        information which consists of the fields listed in the
+                        ipwhois.whois.RIR_WHOIS dictionary.
+                    'raw' (str) - Raw whois results if the inc_raw parameter
+                        is True.
+                    'referral' (dict) - Referral whois information if
+                        get_referral is True and the server is not blacklisted.
+                        Consists of fields listed in the ipwhois.whois.RWHOIS
+                        dictionary.
+                    'raw_referral' (str) - Raw referral whois results if the
+                        inc_raw parameter is True.
+                    'nir' (dict) - ipwhois.nir.NIRWhois() results if inc_nir
+                        is True.
+                }
         """
 
         from .whois import Whois
@@ -215,60 +217,78 @@ class IPWhois:
         information to parse.**
 
         Args:
-            inc_raw: Boolean for whether to include the raw whois results in
-                the returned dictionary.
-            retry_count: The number of times to retry in case socket errors,
-                timeouts, connection resets, etc. are encountered.
-            depth: How many levels deep to run queries when additional
-                referenced objects are found.
-            excluded_entities: A list of entity handles to not perform lookups.
-            bootstrap: If True, performs lookups via ARIN bootstrap rather
-                than lookups based on ASN data. ASN lookups are not performed
-                and no output for any of the asn* fields is provided.
-            rate_limit_timeout: The number of seconds to wait before retrying
-                when a rate limit notice is returned via rdap+json.
-            asn_alts: List of additional lookup types to attempt if the
+            inc_raw (:obj:`bool`): Whether to include the raw whois results in
+                the returned dictionary. Defaults to False.
+            retry_count (:obj:`int`): The number of times to retry in case
+                socket errors, timeouts, connection resets, etc. are
+                encountered. Defaults to 3.
+            depth (:obj:`int`): How many levels deep to run queries when
+                additional referenced objects are found. Defaults to 0.
+            excluded_entities (:obj:`list`): Entity handles to not perform
+                lookups. Defaults to None.
+            bootstrap (:obj:`bool`): If True, performs lookups via ARIN
+                bootstrap rather than lookups based on ASN data. ASN lookups
+                are not performed and no output for any of the asn* fields is
+                provided. Defaults to False.
+            rate_limit_timeout (:obj:`int`): The number of seconds to wait
+                before retrying when a rate limit notice is returned via
+                rdap+json. Defaults to 120.
+            asn_alts (:obj:`list`): Additional lookup types to attempt if the
                 ASN dns lookup fails. Allow permutations must be enabled.
-                Defaults to all ['whois', 'http']. *WARNING* deprecated in
-                favor of new argument asn_methods.
-            extra_org_map: Dictionary mapping org handles to RIRs. This is for
-                limited cases where ARIN REST (ASN fallback HTTP lookup) does
-                not show an RIR as the org handle e.g., DNIC (which is now the
-                built in ORG_MAP) e.g., {'DNIC': 'arin'}. Valid RIR values are
-                (note the case-sensitive - this is meant to match the REST
-                result): 'ARIN', 'RIPE', 'apnic', 'lacnic', 'afrinic'
-            inc_nir: Boolean for whether to retrieve NIR (National Internet
+                If None, defaults to all ['whois', 'http']. *WARNING*
+                deprecated in favor of new argument asn_methods.
+            extra_org_map (:obj:`dict`): Dictionary mapping org handles to
+                RIRs. This is for limited cases where ARIN REST (ASN fallback
+                HTTP lookup) does not show an RIR as the org handle e.g., DNIC
+                (which is now the built in ORG_MAP) e.g., {'DNIC': 'arin'}.
+                Valid RIR values are (note the case-sensitive - this is meant
+                to match the REST result):
+                'ARIN', 'RIPE', 'apnic', 'lacnic', 'afrinic'
+                Defaults to None.
+            inc_nir (:obj:`bool`): Whether to retrieve NIR (National Internet
                 Registry) information, if registry is JPNIC (Japan) or KRNIC
                 (Korea). If True, extra network requests will be required.
                 If False, the information returned for JP or KR IPs is
-                severely restricted.
-            nir_field_list: If provided and inc_nir, a list of fields to parse:
+                severely restricted. Defaults to True.
+            nir_field_list (:obj:`list`): If provided and inc_nir, a list of
+                fields to parse:
                 ['name', 'handle', 'country', 'address', 'postal_code',
                 'nameservers', 'created', 'updated', 'contacts']
-            asn_methods: List of ASN lookup types to attempt, in order.
-                Defaults to all ['dns', 'whois', 'http'].
-            get_asn_description: Boolean for whether to run an additional
+                If None, defaults to all.
+            asn_methods (:obj:`list`): ASN lookup types to attempt, in order.
+                If None, defaults to all ['dns', 'whois', 'http'].
+            get_asn_description (:obj:`bool`): Whether to run an additional
                 query when pulling ASN information via dns, in order to get
-                the ASN description.
+                the ASN description. Defaults to True.
 
         Returns:
-            Dictionary:
+            dict: The IP RDAP lookup results
 
-            :query: The IP address (String)
-            :asn: The Autonomous System Number (String)
-            :asn_date: The ASN Allocation date (String)
-            :asn_registry: The assigned ASN registry (String)
-            :asn_cidr: The assigned ASN CIDR (String)
-            :asn_country_code: The assigned ASN country code (String)
-            :asn_description: The ASN description (String)
-            :entities: List of entity handles referred by the top level query.
-            :network: Dictionary containing network information which consists
-                of the fields listed in the ipwhois.rdap._RDAPNetwork dict.
-            :objects: Dictionary of (entity handle: entity dict) which consists
-                of the fields listed in the ipwhois.rdap._RDAPEntity dict.
-            :raw: (Dictionary) - Whois results in json format if the inc_raw
-                parameter is True.
-            :nir: (Dictionary) - nir.NIRWhois() results if inc_nir is True.
+            ::
+
+                {
+                    'query' (str) - The IP address
+                    'asn' (str) - The Autonomous System Number
+                    'asn_date' (str) - The ASN Allocation date
+                    'asn_registry' (str) - The assigned ASN registry
+                    'asn_cidr' (str) - The assigned ASN CIDR
+                    'asn_country_code' (str) - The assigned ASN country code
+                    'asn_description' (str) - The ASN description
+                    'entities' (list) - Entity handles referred by the top
+                        level query.
+                    'network' (dict) - Network information which consists of
+                        the fields listed in the ipwhois.rdap._RDAPNetwork
+                        dict.
+                    'objects' (dict) - Mapping of entity handle->entity dict
+                        which consists of the fields listed in the
+                        ipwhois.rdap._RDAPEntity dict. The raw result is
+                        included for each object if the inc_raw parameter
+                        is True.
+                    'raw' (dict) - Whois results in json format if the inc_raw
+                        parameter is True.
+                    'nir' (dict) - ipwhois.nir.NIRWhois results if inc_nir is
+                        True.
+                }
         """
 
         from .rdap import RDAP