+++ ipwhois/net.py
@@ -26,6 +26,7 @@ import sys
 import socket
 import dns.resolver
 import json
+from collections import namedtuple
 import logging
 from time import sleep
 
@@ -96,12 +97,15 @@ class Net:
     The class for performing network queries.
 
     Args:
-        address: An IPv4 or IPv6 address in string format.
-        timeout: The default timeout for socket connections in seconds.
-        proxy_opener: The urllib.request.OpenerDirector request for proxy
-            support or None.
-        allow_permutations: Use additional methods if DNS lookups to Cymru
-            fail. *WARNING* deprecated in favor of new argument asn_methods.
+        address (:obj:`str`/:obj:`int`/:obj:`IPv4Address`/:obj:`IPv6Address`):
+            An IPv4 or IPv6 address
+        timeout (:obj:`int`): The default timeout for socket connections in
+            seconds. Defaults to 5.
+        proxy_opener (:obj:`urllib.request.OpenerDirector`): The request for
+            proxy support. Defaults to None.
+        allow_permutations (:obj:`bool`): Allow net.Net() to use additional
+            methods if DNS lookups to Cymru fail. *WARNING* deprecated in
+            favor of new argument asn_methods. Defaults to True.
 
     Raises:
         IPDefinedError: The address provided is defined (does not need to be
@@ -236,7 +240,7 @@ class Net:
         Cymru via port 53 (DNS).
 
         Returns:
-            String: The raw ASN data.
+            str: The raw ASN data.
 
         Raises:
             ASNLookupError: The ASN lookup failed.
@@ -269,10 +273,10 @@ class Net:
         Cymru DNS does not return the ASN Description like Cymru Whois does.
 
         Args:
-            asn: The ASN string (required).
+            asn (:obj:`str`): The AS number (required).
 
         Returns:
-            String: The raw ASN data.
+            str: The raw ASN data.
 
         Raises:
             ASNLookupError: The ASN lookup failed.
@@ -310,11 +314,12 @@ class Net:
         Cymru via port 43/tcp (WHOIS).
 
         Args:
-            retry_count: The number of times to retry in case socket errors,
-                timeouts, connection resets, etc. are encountered.
+            retry_count (:obj:`int`): The number of times to retry in case
+                socket errors, timeouts, connection resets, etc. are
+                encountered. Defaults to 3.
 
         Returns:
-            String: The raw ASN data.
+            str: The raw ASN data.
 
         Raises:
             ASNLookupError: The ASN lookup failed.
@@ -378,11 +383,12 @@ class Net:
         chance fallback call behind ASN DNS & ASN Whois lookups.
 
         Args:
-            retry_count: The number of times to retry in case socket errors,
-                timeouts, connection resets, etc. are encountered.
+            retry_count (:obj:`int`): The number of times to retry in case
+                socket errors, timeouts, connection resets, etc. are
+                encountered. Defaults to 3.
 
         Returns:
-            Dictionary: The ASN data in json format.
+            dict: The ASN data in json format.
 
         Raises:
             ASNLookupError: The ASN lookup failed.
@@ -428,16 +434,17 @@ class Net:
         The function for retrieving CIDR info for an ASN via whois.
 
         Args:
-            asn_registry: The source to run the query against
+            asn_registry (:obj:`str`): The source to run the query against
                 (asn.ASN_ORIGIN_WHOIS).
-            asn: The ASN string (required).
-            retry_count: The number of times to retry in case socket errors,
-                timeouts, connection resets, etc. are encountered.
-            server: An optional server to connect to. Defaults to RADB.
-            port: The network port to connect on.
+            asn (:obj:`str`): The AS number (required).
+            retry_count (:obj:`int`): The number of times to retry in case
+                socket errors, timeouts, connection resets, etc. are
+                encountered. Defaults to 3.
+            server (:obj:`str`): An optional server to connect to.
+            port (:obj:`int`): The network port to connect on. Defaults to 43.
 
         Returns:
-            String: The raw ASN origin whois data.
+            str: The raw ASN origin whois data.
 
         Raises:
             WhoisLookupError: The ASN origin whois lookup failed.
@@ -540,17 +547,19 @@ class Net:
         address via any port. Defaults to port 43/tcp (WHOIS).
 
         Args:
-            asn_registry: The NIC to run the query against.
-            retry_count: The number of times to retry in case socket errors,
-                timeouts, connection resets, etc. are encountered.
-            server: An optional server to connect to. If provided, asn_registry
-                will be ignored.
-            port: The network port to connect on.
-            extra_blacklist: A list of blacklisted whois servers in addition to
-                the global BLACKLIST.
+            asn_registry (:obj:`str`): The NIC to run the query against.
+                Defaults to 'arin'.
+            retry_count (:obj:`int`): The number of times to retry in case
+                socket errors, timeouts, connection resets, etc. are
+                encountered. Defaults to 3.
+            server (:obj:`str`): An optional server to connect to. If
+                provided, asn_registry will be ignored.
+            port (:obj:`int`): The network port to connect on. Defaults to 43.
+            extra_blacklist (:obj:`list` of :obj:`str`): Blacklisted whois
+                servers in addition to the global BLACKLIST. Defaults to None.
 
         Returns:
-            String: The raw whois data.
+            str: The raw whois data.
 
         Raises:
             BlacklistError: Raised if the whois server provided is in the
@@ -666,16 +675,18 @@ class Net:
         The function for retrieving a json result via HTTP.
 
         Args:
-            url: The URL to retrieve.
-            retry_count: The number of times to retry in case socket errors,
-                timeouts, connection resets, etc. are encountered.
-            rate_limit_timeout: The number of seconds to wait before retrying
-                when a rate limit notice is returned via rdap+json.
-            headers: The HTTP headers dictionary. The Accept header defaults
-                to 'application/rdap+json'.
+            url (:obj:`str`): The URL to retrieve (required).
+            retry_count (:obj:`int`): The number of times to retry in case
+                socket errors, timeouts, connection resets, etc. are
+                encountered. Defaults to 3.
+            rate_limit_timeout (:obj:`int`): The number of seconds to wait
+                before retrying when a rate limit notice is returned via
+                rdap+json or HTTP error 429. Defaults to 60.
+            headers (:obj:`dict`): The HTTP headers. The Accept header
+                defaults to 'application/rdap+json'.
 
         Returns:
-            Dictionary: The data in json format.
+            dict: The data in json format.
 
         Raises:
             HTTPLookupError: The HTTP lookup failed.
@@ -795,11 +806,17 @@ class Net:
         The function for retrieving host information for an IP address.
 
         Args:
-            retry_count: The number of times to retry in case socket errors,
-                timeouts, connection resets, etc. are encountered.
+            retry_count (:obj:`int`): The number of times to retry in case
+                socket errors, timeouts, connection resets, etc. are
+                encountered. Defaults to 3.
 
         Returns:
-            Tuple: hostname, aliaslist, ipaddrlist
+            namedtuple:
+
+            :hostname (str): The hostname returned mapped to the given IP
+                address.
+            :aliaslist (list): Alternate names for the given IP address.
+            :ipaddrlist (list): IPv4/v6 addresses mapped to the same hostname.
 
         Raises:
             HostLookupError: The host lookup failed.
@@ -820,7 +837,9 @@ class Net:
 
                 socket.setdefaulttimeout(None)
 
-            return ret
+            results = namedtuple('get_host_results', 'hostname, aliaslist, '
+                                                     'ipaddrlist')
+            return results(ret)
 
         except (socket.timeout, socket.error) as e:
 
@@ -850,16 +869,18 @@ class Net:
         The function for retrieving a raw HTML result via HTTP.
 
         Args:
-            url: The URL to retrieve.
-            retry_count: The number of times to retry in case socket errors,
-                timeouts, connection resets, etc. are encountered.
-            headers: The HTTP headers dictionary. The Accept header defaults
-                to 'text/html'.
-            request_type: 'GET' or 'POST'
-            form_data: Dictionary of form POST data
+            url (:obj:`str`): The URL to retrieve (required).
+            retry_count (:obj:`int`): The number of times to retry in case
+                socket errors, timeouts, connection resets, etc. are
+                encountered. Defaults to 3.
+            headers (:obj:`dict`): The HTTP headers. The Accept header
+                defaults to 'text/html'.
+            request_type (:obj:`str`): Request type 'GET' or 'POST'. Defaults
+                to 'GET'.
+            form_data (:obj:`dict`): Optional form POST data.
 
         Returns:
-            String: The raw data.
+            str: The raw data.
 
         Raises:
             HTTPLookupError: The HTTP lookup failed.